pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spigym"
version = "0.0.1"
description = "SPi-Active: Sampling-based SysID"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "hydra-core>=1.2.0",
  "numpy==1.23.5",
  "rich",
  "ipdb",
  "matplotlib",
  "termcolor",
  "wandb",
  "plotly",
  "tqdm",
  "loguru",
  "meshcat",
  "pynput",
  "scipy",
  "tensorboard",
  "onnx",
  "onnxruntime<1.20.0",
  "opencv-python",
  "joblib",
  "easydict",
  "lxml",
  "numpy-stl",
  "open3d",
  "optuna",
]

[project.urls]
Homepage = "https://github.com/LeCAR-Lab/spi-active"

[tool.setuptools.packages.find]
# mirror setup.py's find_packages()
include = ["*"]

